Refactor dataset and section methods; update tests
Renamed methods (e.g., `add_item` to `add_row`), modified argument usage, and updated database engine handling to enhance clarity and consistency. Revised test cases to align with these changes and improve test coverage.
